Switch nih_main_loop to using epoll
Without this patch, any libnih daemon which opens > 1024 fds can
crash.  This has been known to happen.  With this patch, that limit
is removed.

Two new helpers, nih_io_setup_epoll_fds() and nih_io_handle_epoll_fds(),
are introduced.  These can replace their original counterparts,
nih_io_select_fds() and nih_io_handle_fds().

Callers may continue to use the originals, in which case they continue
to be limited to 1024 fds.  Or they can switch to the new versions, and
not have this limit.  Obviously they must not mix and match within one
iteration, as the output of the first is used as input to the second.